# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Batra Butterfly Footwear Business OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: retailerId
          in: query
          required: false
          schema:
            type: integer
        - name: staffId
          in: query
          required: false
          schema:
            type: integer
        - name: from
          in: query
          required: false
          schema:
            type: string
        - name: to
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceWithItems"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithItems"
  /invoices/daily-summary:
    get:
      operationId: getDailySalesSummary
      tags: [invoices]
      summary: Get daily sales summary
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Daily sales grouped by date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailySalesEntry"
  /invoices/analyze-image:
    post:
      operationId: analyzeInvoiceImage
      tags: [invoices]
      summary: Analyze invoice photo with Claude AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeInvoiceImageBody"
      responses:
        "200":
          description: Extracted invoice data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceAnalysisResult"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithItems"
        "404":
          description: Not found
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete a draft invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /invoices/{id}/confirm:
    post:
      operationId: confirmInvoice
      tags: [invoices]
      summary: Confirm invoice and cascade updates to stock, ledger, and commissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Confirmed invoice with cascade results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceConfirmResult"
        "404":
          description: Not found
...
components:
  schemas:
...
    LedgerEntry:
      type: object
      properties:
        id:
          type: integer
        retailerId:
          type: integer
        type:
          type: string
          enum: [sale, payment]
        amount:
          type: number
        note:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - retailerId
        - type
        - amount
        - note
        - date
        - createdAt
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        retailerId:
          type: integer
        staffId:
          type: integer
        amount:
          type: number
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - retailerId
        - staffId
        - amount
        - date
        - createdAt
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        productId:
          type: ["integer", "null"]
        articleCode:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
      required:
        - id
        - invoiceId
        - productId
        - articleCode
        - productName
        - quantity
        - unitPrice
        - totalPrice
    InvoiceWithItems:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        retailerId:
          type: integer
        retailerName:
          type: string
        staffId:
          type: integer
        staffName:
          type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [draft, confirmed, cancelled]
        imageUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
      required:
        - id
        - invoiceNumber
        - retailerId
        - retailerName
        - staffId
        - staffName
        - totalAmount
        - status
        - imageUrl
        - notes
        - date
        - createdAt
        - items
    CreateInvoiceItemInput:
      type: object
      properties:
        productId:
          type: integer
        articleCode:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required:
        - articleCode
        - productName
        - quantity
        - unitPrice
    CreateInvoiceBody:
      type: object
      properties:
        retailerId:
          type: integer
        staffId:
          type: integer
        invoiceNumber:
          type: string
        date:
          type: string
        notes:
          type: string
        imageUrl:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateInvoiceItemInput"
      required:
        - retailerId
        - staffId
        - invoiceNumber
        - items
    AnalyzeInvoiceImageBody:
      type: object
      properties:
        imageBase64:
          type: string
        mimeType:
          type: string
      required:
        - imageBase64
    InvoiceAnalysisResult:
      type: object
      properties:
        invoiceNumber:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        retailerName:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedInvoiceItem"
        rawText:
          type: string
        confidence:
          type: number
      required:
        - invoiceNumber
        - date
        - retailerName
        - totalAmount
        - items
        - rawText
        - confidence
    ExtractedInvoiceItem:
      type: object
      properties:
        articleCode:
          type: ["string", "null"]
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: ["number", "null"]
        totalPrice:
          type: ["number", "null"]
      required:
        - productName
        - quantity
        - articleCode
        - unitPrice
        - totalPrice
    InvoiceConfirmResult:
      type: object
      properties:
        invoice:
          $ref: "#/components/schemas/InvoiceWithItems"
        stockUpdates:
          type: array
          items:
            type: object
            properties:
              articleCode:
                type: string
              productName:
                type: string
              quantityDeducted:
                type: integer
              newStock:
                type: integer
            required:
              - articleCode
              - productName
              - quantityDeducted
              - newStock
        ledgerEntry:
          $ref: "#/components/schemas/LedgerEntry"
        saleRecord:
          $ref: "#/components/schemas/Sale"
        commissionEarned:
          type: number
      required:
        - invoice
        - stockUpdates
        - ledgerEntry
        - saleRecord
        - commissionEarned
    DailySalesEntry:
      type: object
      properties:
        date:
          type: string
        totalAmount:
          type: number
        invoiceCount:
          type: integer
        topRetailer:
          type: ["string", "null"]
      required:
        - date
        - totalAmount
        - invoiceCount
        - topRetailer
...